view-analysis-improving
Summary:

- Table analysis isolated to support recursion in future.
- Plan entrypoint split apart to support recursion in future.
- Removed dead code.
- Added `TODO` hints.
- Build linux executable statically to mitigate WSL shared object errors.
