+++ internal/stackql/internaldto/output_ctx.go
@@ -1,0 +1,11 @@
+package internaldto
+
+import (
+	"github.com/jeroenrinzema/psql-wire/pkg/sqldata"
+	"github.com/stackql/stackql/internal/stackql/dto"
+)
+
+type OutputContext struct {
+	RuntimeContext dto.RuntimeCtx
+	Result         sqldata.ISQLResultStream
+}
