+++ internal/stackql/internaldto/primitive_ctx.go
@@ -6,29 +6,39 @@ import (
 	"github.com/stackql/stackql/internal/stackql/dto"
 )
 
-type BasicPrimitiveContext struct {
+var (
+	_ BasicPrimitiveContext = &standardBasicPrimitiveContext{}
+)
+
+type BasicPrimitiveContext interface {
+	GetAuthContext(prov string) (*dto.AuthCtx, error)
+	GetErrWriter() io.Writer
+	GetWriter() io.Writer
+}
+
+type standardBasicPrimitiveContext struct {
 	body      map[string]interface{}
 	authCtx   func(string) (*dto.AuthCtx, error)
 	writer    io.Writer
 	errWriter io.Writer
 }
 
-func NewBasicPrimitiveContext(authCtx func(string) (*dto.AuthCtx, error), writer io.Writer, errWriter io.Writer) *BasicPrimitiveContext {
-	return &BasicPrimitiveContext{
+func NewBasicPrimitiveContext(authCtx func(string) (*dto.AuthCtx, error), writer io.Writer, errWriter io.Writer) BasicPrimitiveContext {
+	return &standardBasicPrimitiveContext{
 		authCtx:   authCtx,
 		writer:    writer,
 		errWriter: errWriter,
 	}
 }
 
-func (bpp *BasicPrimitiveContext) GetAuthContext(prov string) (*dto.AuthCtx, error) {
+func (bpp *standardBasicPrimitiveContext) GetAuthContext(prov string) (*dto.AuthCtx, error) {
 	return bpp.authCtx(prov)
 }
 
-func (bpp *BasicPrimitiveContext) GetWriter() io.Writer {
+func (bpp *standardBasicPrimitiveContext) GetWriter() io.Writer {
 	return bpp.writer
 }
 
-func (bpp *BasicPrimitiveContext) GetErrWriter() io.Writer {
+func (bpp *standardBasicPrimitiveContext) GetErrWriter() io.Writer {
 	return bpp.errWriter
 }
