+++ internal/stackql/internaldto/rows_dto.go
@@ -1,0 +1,8 @@
+package internaldto
+
+type RowsDTO struct {
+	RowMap      map[string]map[string]interface{}
+	ColumnOrder []string
+	Err         error
+	RowSort     func(map[string]map[string]interface{}) []string
+}
