+++ internal/stackql/relationaldto/relational_table.go
@@ -4,7 +4,6 @@ import "github.com/stackql/stackql/internal/stackql/internaldto"
 
 var (
 	_ RelationalTable = &standardRelationalTable{}
-	_ RelationalTable = &standardRelationalView{}
 )
 
 type RelationalTable interface {
@@ -12,22 +11,14 @@ type RelationalTable interface {
 	GetBaseName() string
 	GetColumns() []RelationalColumn
 	GetName() (string, error)
-	IsView() bool
+	GetView() (internaldto.ViewDTO, bool)
 	PushBackColumn(RelationalColumn)
 	WithAlias(alias string) RelationalTable
+	WithView(viewDTO internaldto.ViewDTO) RelationalTable
 }
 
 func NewRelationalTable(hIDs internaldto.HeirarchyIdentifiers, discoveryID int, name, baseName string) RelationalTable {
 	return &standardRelationalTable{
-		hIDs:        hIDs,
-		name:        name,
-		baseName:    baseName,
-		discoveryID: discoveryID,
-	}
-}
-
-func NewRelationalView(hIDs internaldto.HeirarchyIdentifiers, discoveryID int, name, baseName string) RelationalTable {
-	return &standardRelationalView{
 		hIDs:        hIDs,
 		name:        name,
 		baseName:    baseName,
