+++ internal/stackql/sqldialect/sql/postgres/sqlengine-setup.ddl
@@ -98,13 +98,11 @@ ON "__iql__.views" (view_name)
 
 INSERT INTO "__iql__.views" (
   view_name,
-  view_ddl,
-  view_stackql_ddl
+  view_ddl
 ) 
 VALUES (
-  'stackql_providers',
-  'SHOW PROVIDERS',
-  'SHOW PROVIDERS'
+  'stackql_repositories',
+  'select id, name, url from github.repos.repos where org = ''stackql'';'
 )
 ON CONFLICT (view_name) DO NOTHING
 ;
